test(e2e): split dcode TUI helper coverage
